requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytils==0.4.1
pytz==2024.1
sqlparse==0.5.0
tzdata==2024.1
django-filter==24.2